+++ CHANGELOG.md
@@ -6,6 +6,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased](https://github.com/alexdlaird/java-ngrok/compare/2.3.4...HEAD)
 
+### Added
+
+- Support for `traffic_policy` tunnel definition, which `ngrok` renamed from `policy`.
+- Support for `traffic_policy.on_http_request` and `traffic_policy.on_http_response`, which `ngrok` added unique keys for in HTTP tunnels.
+
+### Remove
+
+- Tunnel definition fields `hostname` and `client_cas`, as they are no longer supported by `ngrok`.
+
 ## [2.3.4](https://github.com/alexdlaird/java-ngrok/compare/2.3.3...2.3.4) - 2024-12-13
 
 ### Added